shadcn re-init, tsconfig bug fixed
