+++ package.json
@@ -11,8 +11,9 @@
   "dependencies": {
     "@mantine/hooks": "^7.0.2",
     "@next/font": "^13.5.4",
-    "@radix-ui/react-dropdown-menu": "^2.0.5",
+    "@radix-ui/react-dropdown-menu": "^2.0.6",
     "@radix-ui/react-icons": "^1.3.0",
+    "@radix-ui/react-slot": "^1.0.2",
     "@react-three/drei": "9.32.4",
     "@react-three/fiber": "8.8.7",
     "@react-three/postprocessing": "2.6.2",
