+++ tsconfig.json
@@ -19,7 +19,7 @@
     "incremental": true,
     "paths": {
       "@/*": [
-        "./@/*"
+        "./*"
       ]
     },
     "plugins": [
